Add new SPSCLI test suites
