+++ .github/workflows/ci.yml
@@ -15,7 +15,10 @@ jobs:
           args: --config .swiftlint.yml
 
   test:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-14]
     steps:
       - uses: actions/checkout@v4
       - uses: swift-actions/setup-swift@v1
@@ -25,12 +28,17 @@ jobs:
         run: swift build
       - name: Test
         run: swift test --enable-code-coverage
+      - name: Test SPS
+        working-directory: sps
+        run: swift test --enable-code-coverage
       - name: Generate Coverage
+        if: matrix.os == 'ubuntu-latest'
         run: |
           llvm-cov-19 export -format=lcov .build/x86_64-unknown-linux-gnu/debug/FountainCoachPackageTests.xctest -instr-profile .build/x86_64-unknown-linux-gnu/debug/codecov/default.profdata > coverage.lcov
           COVERAGE=$(lcov --summary coverage.lcov 2>/dev/null | grep lines | awk '{print $2}' | sed 's/%//')
           curl -s "https://img.shields.io/badge/Coverage-${COVERAGE}%25-blue" -o coverage-badge.svg
       - name: Upload Coverage
+        if: matrix.os == 'ubuntu-latest'
         uses: actions/upload-artifact@v4
         with:
           name: coverage
