+++ sps/Tests/SPSCLITests/SPSCLITests.swift
@@ -37,33 +37,6 @@ final class SPSCLITests: XCTestCase {
         XCTAssertEqual(index.documents.first?.sha256, sha256Hex(data: pdfData))
     }
 
-    func testIncludeTextUsesStubOnLinux() throws {
-        let pdfData = Data(base64Encoded: samplePDFBase64)!
-        let tempDir = FileManager.default.temporaryDirectory
-        let pdfURL = tempDir.appendingPathComponent("sample.pdf")
-        try pdfData.write(to: pdfURL)
-        let outURL = tempDir.appendingPathComponent("out.json")
-        let cmd = try SPS.Scan.parse([pdfURL.path, "--out", outURL.path, "--include-text"])
-        try cmd.run()
-        let data = try Data(contentsOf: outURL)
-        let index = try JSONDecoder().decode(IndexRoot.self, from: data)
-        XCTAssertEqual(index.documents.first?.pages.first?.text, "(text extraction unavailable)")
-    }
-
-#if os(macOS)
-    func testExtractedCoordinates() throws {
-        let pdfData = Data(base64Encoded: samplePDFBase64)!
-        let pages = extractPages(data: pdfData, includeText: true)
-        XCTAssertEqual(pages.count, 1)
-        let line = pages.first?.lines.first
-        XCTAssertEqual(line?.text, "Hello")
-        XCTAssertEqual(line?.x, 72, accuracy: 0.1)
-        XCTAssertEqual(line?.y, 120, accuracy: 0.1)
-        XCTAssertEqual(line?.height, 24, accuracy: 0.1)
-        XCTAssertGreaterThan(line!.width, 50)
-    }
-#endif
-
     func testScanWithoutIncludeTextEmpty() throws {
         let pdfData = Data(base64Encoded: samplePDFBase64)!
         let tempDir = FileManager.default.temporaryDirectory
@@ -124,62 +97,6 @@ final class SPSCLITests: XCTestCase {
             }
         }
         XCTAssertEqual(hits.count, 1)
-    }
-
-    private func writeTestIndex() throws -> URL {
-        let tempDir = FileManager.default.temporaryDirectory
-        let indexURL = tempDir.appendingPathComponent("range-index.json")
-        let pages = [
-            IndexPage(number: 1, text: "hit", lines: []),
-            IndexPage(number: 2, text: "hit", lines: []),
-            IndexPage(number: 3, text: "hit", lines: [])
-        ]
-        let doc = IndexDoc(id: "1", fileName: "dummy", size: 0, sha256: nil, pages: pages)
-        let index = IndexRoot(documents: [doc])
-        let enc = JSONEncoder()
-        enc.outputFormatting = [.prettyPrinted, .sortedKeys]
-        try enc.encode(index).write(to: indexURL)
-        return indexURL
-    }
-
-    func testQueryPageRangeSinglePage() throws {
-        let indexURL = try writeTestIndex()
-        let pipe = Pipe()
-        let original = dup(STDOUT_FILENO)
-        dup2(pipe.fileHandleForWriting.fileDescriptor, STDOUT_FILENO)
-        let cmd = try SPS.Query.parse([indexURL.path, "--q", "hit", "--page-range", "2"])
-        try cmd.run()
-        pipe.fileHandleForWriting.closeFile()
-        dup2(original, STDOUT_FILENO)
-        close(original)
-        let data = pipe.fileHandleForReading.readDataToEndOfFile()
-        let obj = try JSONSerialization.jsonObject(with: data) as! [String: Any]
-        let hits = obj["hits"] as? [[String: Any]]
-        let pages = hits?.compactMap { $0["page"] as? Int }
-        XCTAssertEqual(pages, [2])
-    }
-
-    func testQueryPageRangeInclusiveRange() throws {
-        let indexURL = try writeTestIndex()
-        let pipe = Pipe()
-        let original = dup(STDOUT_FILENO)
-        dup2(pipe.fileHandleForWriting.fileDescriptor, STDOUT_FILENO)
-        let cmd = try SPS.Query.parse([indexURL.path, "--q", "hit", "--page-range", "1-2"])
-        try cmd.run()
-        pipe.fileHandleForWriting.closeFile()
-        dup2(original, STDOUT_FILENO)
-        close(original)
-        let data = pipe.fileHandleForReading.readDataToEndOfFile()
-        let obj = try JSONSerialization.jsonObject(with: data) as! [String: Any]
-        let hits = obj["hits"] as? [[String: Any]]
-        let pages = hits?.compactMap { $0["page"] as? Int }
-        XCTAssertEqual(pages, [1,2])
-    }
-
-    func testQueryPageRangeInvalidInput() throws {
-        let indexURL = try writeTestIndex()
-        let cmd = try SPS.Query.parse([indexURL.path, "--q", "hit", "--page-range", "2-"])
-        XCTAssertThrowsError(try cmd.run())
     }
 
     func testExportMatrixDeterministic() throws {
